+++ src/KSProject.WebApi/appsettings.Development.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "Default": "Server=localhost,1433;Initial Catalog=KSProjectDb;User ID=sa;Password=123;TrustServerCertificate=true",
+    "Default": "Server=localhost,1433;Initial Catalog=KSProjectDb2;User ID=sa;Password=123;TrustServerCertificate=true",
     "Master": "Server=localhost,1433;Initial Catalog=Master;User ID=sa;Password=123;TrustServerCertificate=true"
   },
   "Serilog": {
